third_party/blink/web_tests/platform/generic/http/tests/security/cross-frame-access-location-get-expected.txt


----- tests for getting window.location's properties -----

PASS: canGet('targetWindow.location') should be 'true' and is.
PASS: toString('targetWindow.location', '') should be '' and is.
Firefox allows access to 'location.toString' but throws an exception when you call it.
PASS: canGet('targetWindow.location.toString') should be 'false' and is.
PASS: accessThrowsException('targetWindow.location.toString') should be 'true' and is.
PASS: canGet('targetWindow.location.assign') should be 'false' and is.
PASS: canGet('targetWindow.location.hash') should be 'false' and is.
PASS: canGet('targetWindow.location.host') should be 'false' and is.
PASS: canGet('targetWindow.location.hostname') should be 'false' and is.
PASS: canGet('targetWindow.location.href') should be 'false' and is.
PASS: canGet('targetWindow.location.pathname') should be 'false' and is.
PASS: canGet('targetWindow.location.port') should be 'false' and is.
PASS: canGet('targetWindow.location.protocol') should be 'false' and is.
PASS: canGet('targetWindow.location.reload') should be 'false' and is.
PASS: canGet('targetWindow.location.search') should be 'false' and is.
PASS: canGet('targetWindow.location.existingCustomProperty') should be 'false' and is.
PASS: canGet('targetWindow.location[1]') should be 'false' and is.
PASS: accessThrowsException('targetWindow.location.href') should be 'true' and is.
PASS: accessThrowsException('targetWindow.location.hash') should be 'true' and is.
PASS: accessThrowsException('targetWindow.location.host') should be 'true' and is.
PASS: accessThrowsException('targetWindow.location.hostname') should be 'true' and is.
PASS: accessThrowsException('targetWindow.location.pathname') should be 'true' and is.
PASS: accessThrowsException('targetWindow.location.port') should be 'true' and is.
PASS: accessThrowsException('targetWindow.location.protocol') should be 'true' and is.
PASS: accessThrowsException('targetWindow.location.search') should be 'true' and is.
PASS: accessThrowsException('targetWindow.location.existingCustomProperty') should be 'true' and is.
PASS: accessThrowsException('targetWindow.location[1]') should be 'true' and is.
PASS: canGet('targetWindow.location.replace') should be 'true' and is.
